+++ cms/src/features/login/Login.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 
+import NavAnchor from "../../components/navAnchor/NavAnchor";
 import LoginForm from "./components/LoginForm";
 
 const SignIn: React.FC = () => {
@@ -7,6 +8,10 @@ const SignIn: React.FC = () => {
     <section>
       <h2>Sign in</h2>
       <LoginForm />
+      <p>
+        <span>New here? </span>
+        <NavAnchor pathname="/signup" textContent="Sign up" />
+      </p>
     </section>
   );
 };